// microshop-ui/src/app/api/basket/model/shopping-cart.ts
/**
 * MicroShop.BasketAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BasketItem } from './basket-item';


export interface ShoppingCart { 
    userId?: string;
    items?: Array<BasketItem> | null;
    readonly totalPrice?: number;
}